Fixed serialization issues, removed unnecessary code and reduced duplicate code
