+++ app/src/androidTest/java/de/westnordost/streetcomplete/data/osmnotes/CreateNoteDaoTest.java
@@ -4,6 +4,7 @@ import de.westnordost.streetcomplete.data.ApplicationDbTestCase;
 import de.westnordost.osmapi.map.data.BoundingBox;
 import de.westnordost.osmapi.map.data.Element;
 import de.westnordost.osmapi.map.data.OsmLatLon;
+import de.westnordost.streetcomplete.util.Serializer;
 
 public class CreateNoteDaoTest extends ApplicationDbTestCase
 {
@@ -12,7 +13,7 @@ public class CreateNoteDaoTest extends ApplicationDbTestCase
 	@Override public void setUp()
 	{
 		super.setUp();
-		dao = new CreateNoteDao(dbHelper);
+		dao = new CreateNoteDao(dbHelper, serializer);
 	}
 
 	public void testAddGetDelete()
