+++ app/src/main/java/de/westnordost/streetcomplete/quests/LeaveNoteDialog.java
@@ -3,37 +3,16 @@ package de.westnordost.streetcomplete.quests;
 import android.app.Activity;
 import android.app.DialogFragment;
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.GridView;
 
-import java.io.File;
-import java.io.IOException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 
 import de.westnordost.streetcomplete.R;
-import de.westnordost.streetcomplete.view.NoteImageAdapter;
-import de.westnordost.streetcomplete.view.dialogs.AlertDialogBuilder;
-
-import static android.app.Activity.RESULT_OK;
-import static android.support.v4.content.FileProvider.getUriForFile;
 
 public class LeaveNoteDialog extends DialogFragment
 {
@@ -44,16 +23,9 @@ public class LeaveNoteDialog extends DialogFragment
 
 	private QuestAnswerComponent questAnswerComponent;
 
+	public static ArrayList<String> imagePaths = new ArrayList<>();
+
 	private String questTitle;
-
-	private ArrayList<String> imagePaths = new ArrayList<>();
-	private ArrayList<Bitmap> imageBitmaps = new ArrayList<>();
-
-	File photoFile = null;
-
-	static final int REQUEST_TAKE_PHOTO = 1;
-
-	GridView gridView;
 
 	public LeaveNoteDialog()
 	{
@@ -83,43 +55,6 @@ public class LeaveNoteDialog extends DialogFragment
 			public void onClick(View v)
 			{
 				onClickOk();
-			}
-		});
-		Button takePhoto = view.findViewById(R.id.buttonTakeImage);
-		takePhoto.setOnClickListener(new View.OnClickListener()
-		{
-			@Override public void onClick(View v)
-			{
-				takePhoto();
-			}
-		});
-
-		if (!getActivity().getPackageManager().hasSystemFeature(PackageManager.FEATURE_CAMERA))
-		{
-			takePhoto.setVisibility(View.GONE);
-		}
-
-		final NoteImageAdapter noteImageAdapter = new NoteImageAdapter(getActivity(), imageBitmaps);
-		gridView = view.findViewById(R.id.gridView);
-		gridView.setAdapter(noteImageAdapter);
-		gridView.setOnItemClickListener(new AdapterView.OnItemClickListener()
-		{
-			@Override
-			public void onItemClick(AdapterView<?> parent, View view, final int position, long id)
-			{
-				new AlertDialogBuilder(getActivity())
-						.setMessage(R.string.quest_leave_new_note_photo_delete_title)
-						.setNegativeButton(android.R.string.cancel, null)
-						.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener()
-						{
-							@Override public void onClick(DialogInterface dialog, int which)
-							{
-								imageBitmaps.remove(position);
-								imagePaths.remove(position);
-								noteImageAdapter.notifyDataSetChanged();
-							}
-						})
-						.show();
 			}
 		});
 
@@ -158,63 +93,16 @@ public class LeaveNoteDialog extends DialogFragment
 			noteInput.setError(getResources().getString(R.string.quest_generic_error_field_empty));
 			return;
 		}
-
 		questAnswerComponent.onLeaveNote(questTitle, inputText, imagePaths);
 		dismiss();
 	}
 
 	private void onClickCancel()
 	{
+		AttachPhotoFragment attachPhotoFragment = new AttachPhotoFragment();
+		attachPhotoFragment.deleteImages(imagePaths);
+
 		questAnswerComponent.onSkippedQuest();
 		dismiss();
-	}
-
-	private void takePhoto()
-	{
-		Intent takePhotoIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-		if (takePhotoIntent.resolveActivity(getActivity().getPackageManager()) != null) {
-			try {
-				photoFile = createImageFile();
-			} catch (IOException e) {
-			}
-			if (photoFile != null) {
-				Log.d("photoFile", photoFile.toString());
-				if (Build.VERSION.SDK_INT > 21) {
-					takePhotoIntent.putExtra(MediaStore.EXTRA_OUTPUT, getUriForFile(getActivity(), "de.westnordost.streetcomplete.fileprovider", photoFile));
-				} else {
-					takePhotoIntent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(photoFile));
-				}
-				startActivityForResult(takePhotoIntent, REQUEST_TAKE_PHOTO);
-			}
-		}
-	}
-
-	@Override
-	public void onActivityResult(int requestCode, int resultCode, Intent data) {
-		if (requestCode == REQUEST_TAKE_PHOTO) {
-			if (resultCode == RESULT_OK)
-			{
-				imageBitmaps.add(BitmapFactory.decodeFile(photoFile.getAbsolutePath()));
-				imagePaths.add(photoFile.toString());
-			} else
-			{
-				if (photoFile.exists())
-				{
-					photoFile.delete();
-				}
-			}
-		}
-	}
-
-	private File createImageFile() throws IOException
-	{
-		String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
-		String imageFileName = "JPEG_" + timeStamp + "_";
-		File directory = getActivity().getExternalFilesDir(Environment.DIRECTORY_PICTURES);
-		return File.createTempFile(
-				imageFileName,
-				".jpg",
-				directory
-		);
 	}
 }
