+++ app/src/main/res/layout/quest_note_discussion.xml
@@ -47,19 +47,9 @@
         android:maxLines="12"
         />
 
-    <GridView
-        android:id="@+id/gridView"
+    <fragment
+        android:name="de.westnordost.streetcomplete.quests.AttachPhotoFragment"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:numColumns="3"
-        android:verticalSpacing="10dp" />
-
-    <Button
-        android:id="@+id/buttonTakeImage"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        style="?android:attr/buttonBarButtonStyle"
-        android:layout_gravity="center_horizontal"
-        android:text="@string/quest_leave_new_note_photo" />
+        android:layout_height="wrap_content" />
 
 </merge>
